--- backend/src.old/modules/account/account.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AccountService } from './account.service';
-
-describe('AccountService', () => {
-  let service: AccountService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AccountService],
-    }).compile();
-
-    service = module.get<AccountService>(AccountService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
