--- backend/src.old/modules/user/user.module.ts
@@ -1,14 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-
-@Module({
-  // このモジュールに含まれるサービス
-  providers: [UserService],
-
-  // このモジュールが公開するサービス
-  exports: [UserService],
-
-  // 依存モジュール
-  imports: []
-})
-export class UserModule {}
