+++ backend/src/entities/AccountEntity.ts
@@ -1,2 +1,10 @@
-export type AccountEntity = {
-};
+import { UserEntity } from './UserEntity';
+
+export class AccountEntity {
+  constructor(
+    public accountId: string,
+    public name: string,
+    public passwordAuthEnabled: boolean,
+    public users: UserEntity[]
+  ) {}
+}
