+++ backend/src/entities/UserEntity.ts
@@ -1,2 +1,7 @@
-export type UserEntity = {
-};
+export class UserEntity {
+  constructor(
+    public userId: string,
+    public name: string,
+    public displayName: string
+  ) {}
+}
