--- backend/src/services/PasswordAuthService.ts
@@ -1,8 +1,0 @@
-import { Container, inject, injectable } from 'inversify';
-import { TYPES } from '../container/types';
-import { DatabaseService } from './DatabaseService';
-
-@injectable()
-export class PasswordAuthService {
-  constructor() {}
-}
